armbian-next: avoid writing to disk during configuration; `ANSI_COLOR=none` logging; make CONFIG_DEFS_ONLY=yes runnable without sudo
- when `CONFIG_DEFS_ONLY=yes`, avoid writing the config summary output.log file.
  - refactor that into a function as to be easy to if-out-of
  - don't write to disk during aggregate_content() if `CONFIG_DEFS_ONLY=yes`
  - don't write to disk during show_checklist_variables() if `CONFIG_DEFS_ONLY=yes`
  - don't write to disk during write_deboostrap_list_debug_log() if `CONFIG_DEFS_ONLY=yes`
  - don't compress and rotate logs if `CONFIG_DEFS_ONLY=yes`
- don't pretend to be handling errors we can't handle during var capture
- I foresee a world we can build all .debs without sudo
- and a some kind of split of codebase entrypoint due to that future feature
- some python info.py enhancements, not ready yet

+++ lib/functions/main/config-prepare.sh
@@ -17,23 +17,25 @@ function prepare_and_config_main_build_single() {
 	[[ -z $LANGUAGE ]] && export LANGUAGE="en_US:en"      # set to english if not set
 	[[ -z $CONSOLE_CHAR ]] && export CONSOLE_CHAR="UTF-8" # set console to UTF-8 if not set
 
-	# set log path
-	LOG_SUBPATH=${LOG_SUBPATH:=debug}
-	mkdir -p "${DEST}/${LOG_SUBPATH}" # This creates the logging output.
+	if [[ "${CONFIG_DEFS_ONLY}" != "yes" ]]; then
+		# set log path
+		LOG_SUBPATH=${LOG_SUBPATH:=debug}
+		mkdir -p "${DEST}/${LOG_SUBPATH}" # This creates the logging output.
 
-	# compress and remove old logs, if they exist.
-	if [[ -f "${DEST}/${LOG_SUBPATH}/timestamp" ]]; then
-		if ls "${DEST}/${LOG_SUBPATH}/"*.log &> /dev/null; then
-			display_alert "Archiving previous build logs..." "${DEST}/${LOG_SUBPATH}" "info"
-			(cd "${DEST}/${LOG_SUBPATH}" && tar -czf logs-"$(< timestamp)".tgz ./*.log) # > /dev/null 2>&1
-			rm -f "${DEST}/${LOG_SUBPATH}"/*.log
+		# compress and remove old logs, if they exist.
+		if [[ -f "${DEST}/${LOG_SUBPATH}/timestamp" ]]; then
+			if ls "${DEST}/${LOG_SUBPATH}/"*.log &> /dev/null; then
+				display_alert "Archiving previous build logs..." "${DEST}/${LOG_SUBPATH}" "info"
+				(cd "${DEST}/${LOG_SUBPATH}" && tar -czf logs-"$(< timestamp)".tgz ./*.log) # > /dev/null 2>&1
+				rm -f "${DEST}/${LOG_SUBPATH}"/*.log
+			fi
+			# delete compressed logs older than 7 days
+			find "${DEST}"/${LOG_SUBPATH} -name '*.tgz' -mtime +7 -delete
 		fi
-		# delete compressed logs older than 7 days
-		find "${DEST}"/${LOG_SUBPATH} -name '*.tgz' -mtime +7 -delete
-	fi
 
-	# Mark a timestamp, for next build.
-	date +"%d_%m_%Y-%H_%M_%S" > "${DEST}"/${LOG_SUBPATH}/timestamp
+		# Mark a timestamp, for next build.
+		date +"%d_%m_%Y-%H_%M_%S" > "${DEST}"/${LOG_SUBPATH}/timestamp
+	fi
 
 	# PROGRESS_LOG_TO_FILE is either yes, or unset. (@TODO: this is still used in buildpkg)
 	if [[ $PROGRESS_LOG_TO_FILE != yes ]]; then unset PROGRESS_LOG_TO_FILE; fi
